Big changes on indexing and code cleansing
These indexing changes consist basically on implementing the logic in which we have been working the last few weeks. We changed the indexing logic so that we "freed" the API of a lot of work. Basically the idea here is that we merged all the documents that data and front asked for into a big (and chunky) TSV, so the API doesn't have to process or merge anything, it just has to read it straightforward.

Changes on code cleansing were basically suggested changes by SolarLint to improve code readability and decrease complexity.